Initial CommInit setup
CommInit has been initialized to create a client for each GPU. We maintain a stack of `available` indices into `gpuInfos` to track which GPUs are available. We maintain the invariant that a GPU can only appear in one communicator. While we do not currently support closing a communicator, this stack would allow us to recover GPUs upon closing a communicator. Furthermore, each communicator knows the ranks of the GPUs tracked by the communicator through the `using` array.